+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Maintenance/TradingAccountCreateVM.cs
@@ -100,5 +100,12 @@ namespace TD.WPF.ViewModels.Maintenance
                 }
             }
         }
+
+        public bool ValidateModel()
+        {
+            return ValidateProperty(_tradingAccount, nameof(Title))
+                && ValidateProperty(_tradingAccount, nameof(ExchangeId))
+                && ValidateProperty(_tradingAccount, nameof(DisplayOrder));
+        }
     }
 }